# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/health-check": {
            "get": {
                "description": "Check if the API server is running",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Health Check"
                ],
                "summary": "Check API health",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/users": {
            "get": {
                "description": "Sends a GET request to get user by paging",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get user by paging with a Get request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "page",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "limit",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "sort",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "employeeCode",
                        "name": "employeeCode",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "userFullname",
                        "name": "userFullname",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "userLogin",
                        "name": "userLogin",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rename-service-name-here_internal_model.DefaultPayload-rename-service-name-here_internal_model_GetUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rename-service-name-here_internal_util.AppErrors"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rename-service-name-here_internal_util.AppErrors"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rename-service-name-here_internal_util.AppErrors"
                        }
                    }
                }
            },
            "post": {
                "description": "Sends a POST request to create User",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Create User with a POST request",
                "parameters": [
                    {
                        "description": "User Request",
                        "name": "userRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rename-service-name-here_internal_model.DefaultPayload-rename-service-name-here_internal_model_NewUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rename-service-name-here_internal_model.DefaultPayload-rename-service-name-here_internal_model_NewUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rename-service-name-here_internal_util.AppErrors"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/rename-service-name-here_internal_util.AppErrors"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rename-service-name-here_internal_util.AppErrors"
                        }
                    }
                }
            },
            "patch": {
                "description": "Sends a PATCH request to Update User",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Update User with a PATCH request",
                "parameters": [
                    {
                        "description": "User Request",
                        "name": "userRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rename-service-name-here_internal_model.DefaultPayload-rename-service-name-here_internal_model_NewUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rename-service-name-here_internal_model.DefaultPayload-rename-service-name-here_internal_model_GetUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rename-service-name-here_internal_util.AppErrors"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/rename-service-name-here_internal_util.AppErrors"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rename-service-name-here_internal_util.AppErrors"
                        }
                    }
                }
            }
        },
        "/v1/users/{id}": {
            "get": {
                "description": "Sends a GET request to get user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get user with a Get request",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rename-service-name-here_internal_model.DefaultPayload-rename-service-name-here_internal_model_GetUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rename-service-name-here_internal_util.AppErrors"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rename-service-name-here_internal_util.AppErrors"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rename-service-name-here_internal_util.AppErrors"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "rename-service-name-here_internal_model.DefaultPayload-rename-service-name-here_internal_model_GetUserResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/rename-service-name-here_internal_model.GetUserResponse"
                }
            }
        },
        "rename-service-name-here_internal_model.DefaultPayload-rename-service-name-here_internal_model_NewUserRequest": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/rename-service-name-here_internal_model.NewUserRequest"
                }
            }
        },
        "rename-service-name-here_internal_model.DefaultPayload-rename-service-name-here_internal_model_NewUserResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/rename-service-name-here_internal_model.NewUserResponse"
                }
            }
        },
        "rename-service-name-here_internal_model.GetUserResponse": {
            "description": "Detailed description for get User response",
            "type": "object",
            "properties": {
                "createdDate": {
                    "type": "string"
                },
                "createdUserId": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "employeeCode": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isActive": {
                    "type": "boolean"
                },
                "nameEnglish": {
                    "type": "string"
                },
                "nameThai": {
                    "type": "string"
                },
                "surnameEnglish": {
                    "type": "string"
                },
                "surnameThai": {
                    "type": "string"
                },
                "updatedDate": {
                    "type": "string"
                },
                "updatedUserId": {
                    "type": "integer"
                },
                "userLogin": {
                    "type": "string"
                }
            }
        },
        "rename-service-name-here_internal_model.NewUserRequest": {
            "description": "Detailed description for new User request",
            "type": "object",
            "required": [
                "email",
                "employeeCode",
                "nameThai",
                "surnameThai",
                "userLogin"
            ],
            "properties": {
                "createdUserId": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "employeeCode": {
                    "type": "string"
                },
                "isActive": {
                    "type": "boolean"
                },
                "nameEnglish": {
                    "type": "string"
                },
                "nameThai": {
                    "type": "string"
                },
                "surnameEnglish": {
                    "type": "string"
                },
                "surnameThai": {
                    "type": "string"
                },
                "userLogin": {
                    "type": "string"
                }
            }
        },
        "rename-service-name-here_internal_model.NewUserResponse": {
            "description": "Detailed description for new User response",
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "rename-service-name-here_internal_util.AppErrors": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rename-service-name-here_internal_util.ItemError"
                    }
                }
            }
        },
        "rename-service-name-here_internal_util.ItemError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        }
    }
}